% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predictions.R
\name{get_predictions}
\alias{get_predictions}
\title{Get Predictions}
\usage{
get_predictions(sinusfit_sw, sinusfit_gw, retardation_factor = 2)
}
\arguments{
\item{sinusfit_sw}{as retrieved by optimise_sinus_variablePeriod(temp_df = temp_surfacewater)}

\item{sinusfit_gw}{as retrieved by optimise_sinus_variablePeriod(temp_df = temp_groundwater)}

\item{retardation_factor}{hydraulic retardation factor (default: 2)}
}
\value{
list with sim/observation data ("data") fit parameters ("paras"), goodness-of-fit values ("gof")
and traveltimes ("traveltimes")
}
\description{
Get Predictions
}
