% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_travel_time.R
\name{get_travel_time}
\alias{get_travel_time}
\title{Helper function: get traveltime}
\usage{
get_travel_time(sinusfit_sw, sinusfit_gw, retardation_factor = 1.8)
}
\arguments{
\item{sinusfit_sw}{as retrieved by \code{\link{optimise_sinus_variablePeriod}} with surface water temperature data}

\item{sinusfit_gw}{as retrieved by \code{\link{optimise_sinus_variablePeriod}} with groundwater temperature data}

\item{retardation_factor}{hydraulic retardation factor (default: 2)}
}
\value{
data frame with travel times for min/max and turning points
}
\description{
Helper function: get traveltime
}
