% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_temperature_from_csv.R
\name{load_temperature_from_csv}
\alias{load_temperature_from_csv}
\title{Load Temperature Data From CSV}
\usage{
load_temperature_from_csv(path)
}
\arguments{
\item{path}{path to csv file with temperature data and columns: "date" (YYYY-MM-DD) and "value"}
}
\value{
tibble with temperature data and columns "date" and "value"
}
\description{
Load Temperature Data From CSV
}
\examples{
path <- kwb.heatsine::extdata_file("temperature_groundwater_TEG343.csv")
gw_data <- kwb.heatsine::load_temperature_from_csv(path)
gw_data

}
