% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prediction_interactive.R
\name{mean_cl_quantile}
\alias{mean_cl_quantile}
\title{Helper function: mean_cl_quantile}
\usage{
mean_cl_quantile(x, q = c(0.1, 0.9), na.rm = TRUE)
}
\arguments{
\item{x}{x values}

\item{q}{quantiles (default: c(0.1,0.9))}

\item{na.rm}{(default: TRUE)}
}
\value{
data frame with y, ymin, ymax
}
\description{
Helper function: mean_cl_quantile
}
\keyword{internal}
