% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise_sinus.R
\name{optimise_sinus_fixedPeriod}
\alias{optimise_sinus_fixedPeriod}
\title{Optimise Sinus Fit for Fixed Period}
\usage{
optimise_sinus_fixedPeriod(df, period_length = 365.25)
}
\arguments{
\item{df}{data frame with temperature data and columns "type" (i.e. "groundwater", "surface-water"), "date" and "value"}

\item{period_length}{period length (default: 365.25)}
}
\value{
list with fit parameters ("paras"), goodness-of-fit values ("gof"), special points, i.e. min/max/turning-points
("points), fit model ("lm_model") and input data ("data")
}
\description{
Optimise Sinus Fit for Fixed Period
}
\references{
https://stats.stackexchange.com/questions/77543/how-do-i-get-the-amplitude-and-phase-for-sine-wave-from-lm-summary
}
