% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prediction_interactive.R
\name{plot_prediction_interactive}
\alias{plot_prediction_interactive}
\title{Plot Prediction Interactive}
\usage{
plot_prediction_interactive(predictions)
}
\arguments{
\item{predictions}{as retrieved by get_predictions()}
}
\value{
interactive prediction plot
}
\description{
Plot Prediction Interactive
}
