% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_temperature_interactive.R
\name{plot_temperature_interactive}
\alias{plot_temperature_interactive}
\title{Plot Temperature Interactive}
\usage{
plot_temperature_interactive(df)
}
\arguments{
\item{df}{data frame with temperature data and columns "date" (YYYY-MM-DD) and "value"}
}
\value{
plot with interactive temperature data
}
\description{
Plot Temperature Interactive
}
\examples{
path <- kwb.heatsine::extdata_file("temperature_groundwater_TEG343.csv")
gw_data <- kwb.heatsine::load_temperature_from_csv(path)
kwb.heatsine::plot_temperature_interactive(gw_data)

}
