% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_timeperiod.R
\name{select_timeperiod}
\alias{select_timeperiod}
\title{Helper function: select timeperiod}
\usage{
select_timeperiod(
  df,
  date_start,
  date_end = as.Date(date_start) + 365.25,
  col_date = "date"
)
}
\arguments{
\item{df}{data frame with date (defined in parameter "col_date)}

\item{date_start}{start date of selection}

\item{date_end}{end date of selection. If no value is given 365.25 days after "date_start" will be used
(default: as.Date(date_start) + 365.25)}

\item{col_date}{column for dates (default: "date")}
}
\value{
data frame with selected dates
}
\description{
Helper function: select timeperiod
}
\examples{
path <- kwb.heatsine::extdata_file("temperature_groundwater_Txxxx3.csv")
gw_data <- kwb.heatsine::load_temperature_from_csv(path)
gw_data_selected <- kwb.heatsine::select_timeperiod(gw_data,
date_start = "2015-12-28", date_end = "2016-12-26")
gw_data_selected

}
