% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predictions.R
\name{get_predictions}
\alias{get_predictions}
\title{Get Predictions}
\usage{
get_predictions(sinusfit_sw, sinusfit_gw, retardation_factor = 2)
}
\arguments{
\item{sinusfit_sw}{as retrieved by \code{\link{optimise_sinus_variablePeriod}} with surface water temperature data}

\item{sinusfit_gw}{as retrieved by \code{\link{optimise_sinus_variablePeriod}} with groundwater temperature data}

\item{retardation_factor}{hydraulic retardation factor (default: 2)}
}
\value{
list with sim/observation data ("data") fit parameters ("paras"), goodness-of-fit values ("gof")
traveltimes ("traveltimes") and special (min, max, turning) points ("points")
}
\description{
Get Predictions
}
