% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise_sinus.R
\name{opt_func}
\alias{opt_func}
\title{Optimise Sinus Fit Function}
\usage{
opt_func(period, df, opt_criteria = "RMSE", debug = TRUE)
}
\arguments{
\item{period}{period length}

\item{df}{data frame with temperature data and columns "date" (YYYY-MM-DD)
and "value"}

\item{opt_criteria}{(default: "RMSE"), for other options check:
?hydroGOF::gof}

\item{debug}{show debug messages (default: TRUE)}
}
\value{
scalar with optimisation result
}
\description{
Optimise Sinus Fit Function
}
