% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_temperature_interactive.R
\name{plot_temperature_interactive}
\alias{plot_temperature_interactive}
\title{Plot Temperature Interactive}
\usage{
plot_temperature_interactive(df)
}
\arguments{
\item{df}{data frame with temperature data and columns "type" (i.e. "groundwater", "surface-water"), "date" and "value"}
}
\value{
plot with interactive temperature data
}
\description{
Plot Temperature Interactive
}
